coordinator: fix typos and small code cosmetics
