+++ jvm-libs/generic/json-rpc/src/main/kotlin/net/consensys/linea/jsonrpc/httpserver/HttpJsonRpcServer.kt
@@ -19,7 +19,7 @@ class HttpJsonRpcServer(
 ) : AbstractVerticle() {
   private val log: Logger = LogManager.getLogger(this.javaClass)
   private lateinit var httpServer: HttpServer
-  val bindedPort: Int
+  val boundPort: Int
     get() = if (this::httpServer.isInitialized) {
       httpServer.actualPort()
     } else {
